+++ src/form/field/SelectField.tsx
@@ -1,31 +1,29 @@
 import React from "react";
-import { Field, FieldProps } from "formik";
 import { Select, ISelectProps } from "../../control/Select";
+import { CustomField } from "./CustomField";
 
-export interface ISelectFieldProps extends ISelectProps {
-    id?: string;
+export interface ISelectFieldProps extends Omit<ISelectProps, "value" | "onSelect"> {
     name: string;
-    value?: string;
-    disabled?: boolean;
     validate?(value: string): string | Promise<string | void> | undefined;
 }
 
 export class SelectField extends React.Component<ISelectFieldProps> {
     render() {
-        // We pick and discard some props in order to pass the rest to the inner Input
+        // We pick and discard some props in order to pass the rest to the inner Select
         // because formik doesn't expose them in the children callback
-        let { validate: _, ...inputProps } = this.props;
+        let { validate, name, ...selectProps } = this.props;
 
-        return <Field {...this.props}>
-                {({ field }: FieldProps<any>) => {
-                    return <Select {...inputProps} {...field}
-                        onSelect={(value: string) => {
-                            const fakeEvent = { target: { name: field.name, value }};
-                            field.onChange(fakeEvent);
-                            setTimeout(() => field.onBlur(fakeEvent));
+        return <CustomField name={name} validate={validate}>
+                {({ value: currentValue, onChange, onBlur }) => {
+                    return <Select
+                        {...selectProps}
+                        value={currentValue}
+                        onSelect={value => {
+                            onChange(value);
+                            onBlur();
                         }}
                     />;
                 }}
-            </Field>;
+            </CustomField>;
     }
 }